+++ dev-tools/cert-dev/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "cert-dev"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow.workspace = true
+camino.workspace = true
+clap.workspace = true
+libc.workspace = true
+omicron-workspace-hack.workspace = true
+tokio.workspace = true
+rcgen.workspace = true
+
+[dev-dependencies]
+camino-tempfile.workspace = true
+omicron-test-utils.workspace = true
+openssl.workspace = true
+subprocess.workspace = true
