+++ dev-tools/ch-dev/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "omicron-dev"
+name = "ch-dev"
 version = "0.1.0"
 edition = "2021"
 license = "MPL-2.0"
@@ -16,21 +16,9 @@ clap.workspace = true
 dropshot.workspace = true
 futures.workspace = true
 libc.workspace = true
-nexus-config.workspace = true
-nexus-test-interface.workspace = true
-nexus-test-utils = { workspace = true, features = ["omicron-dev"] }
-omicron-nexus.workspace = true
+omicron-test-utils.workspace = true
 omicron-workspace-hack.workspace = true
 # See omicron-rpaths for more about the "pq-sys" dependency.
 pq-sys = "*"
 signal-hook-tokio.workspace = true
 tokio.workspace = true
-toml.workspace = true
-
-[dev-dependencies]
-expectorate.workspace = true
-omicron-dev-lib.workspace = true
-omicron-test-utils.workspace = true
-oxide-client.workspace = true
-subprocess.workspace = true
-tokio-postgres.workspace = true
