+++ dev-tools/db-dev/tests/output/cmd-db-dev-populate-noargs-stderr
@@ -1,6 +1,6 @@
 error: the following required arguments were not provided:
   --database-url <DATABASE_URL>
 
-Usage: omicron-dev db-populate --database-url <DATABASE_URL>
+Usage: db-dev populate --database-url <DATABASE_URL>
 
 For more information, try '--help'.
