+++ dev-tools/omicron-dev-lib/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "omicron-dev-lib"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+libc.workspace = true
+omicron-test-utils.workspace = true
+omicron-workspace-hack.workspace = true
+subprocess.workspace = true
