+++ dev-tools/omicron-dev-lib/src/lib.rs
@@ -1,0 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Common code shared by the various -dev tools.
+
+pub mod test_utils;
