+++ docs/repo.adoc
@@ -15,7 +15,7 @@ To help the development process, we seek:
 
 * to have clear and up-to-date https://github.com/oxidecomputer/omicron/blob/main/docs/how-to-run-simulated.adoc#installing-prerequisites[instructions] for setting up a development environment from scratch.  Most of this is automated.  CI uses the same automation to go from a bare environment to one that builds and tests Omicron, so this automation is tested regularly with the rest of the repo.
 * to have clear instructions for basic activities like formatting code, running clippy, running tests, etc.  These should be consistent across components and across local development vs. CI.
-* to prioritize debugging and fixing flaky tests so that developers can always expect the tests to pass.  Failures don't necessarily need to be reproducible to debug them.  The test suite preserves trace-level log files and database contents from failed test runs.  You can inspect the database contents using `omicron-dev db-run` to spin up a transient database instance pointed at the saved database contents.
+* to prioritize debugging and fixing flaky tests so that developers can always expect the tests to pass.  Failures don't necessarily need to be reproducible to debug them.  The test suite preserves trace-level log files and database contents from failed test runs.  You can inspect the database contents using `db-dev run` to spin up a transient database instance pointed at the saved database contents.
 * to ensure that a fresh clone and build of the repo should produce equivalent software to any other clone, including the CI environment.  If tests pass for one developer on the tip of "main", they should pass for other developers as well as CI.  We use rust-toolchain and Cargo.lock to ensure that developers are getting a consistent toolchain and packages as each other and CI.
 
 Omicron houses many related components in one repo:
